This PR removes the `--api-key` option from `login` command. Instead of using command-line option to add the API key, the CLI will ask it.
I wanted to get started with a simple use case from prompting user input, but the possibilities how to use this are endless. We could for example:

- Fetch list of marketplaces the user has access and let the user select which marketplace to use
- Fetch list of processes and version for `process pull` command and let the user select which marketplace to use

We could even hook the prompt together with the command definitions so that if in the command definition we define an option as required, it will be automatically prompted if it's not given. However, that requires more thinking because it's not always the individual option that is required (for example, process can be pulled by version OR alias and one of they has to be present)

**Question that arose during implementation**

- Q: By default inquirer outputs to `stdout`. Is that the right place?
  - A: I guess `stderr` is the right place. In case there's a command that prompts user input but user wants to direct the result to a file, `stdout` should not be used.
- Q: Should we throw an error if we try to prompt in a non TTY environment?
  - A: I guess not. If user wants to direct the result to a file, it makes the out stream a non TTY stream. And to me it sounds useful to be able to prompt but then direct output to file.