Improve async error handling
Currently many of our errors end up to the uncaught exception handler due to the fact that they are thrown inside an async function (i.e. in go blocks and channels). The uncaught exception handler is able to format the exception properly and in production mode exit the program with non-zero exit code. In dev environment we don't exit the program because we want to keep it running to eval code from REPL.

This causes couple problems:

- In general it's a bit border line bad practice to rely on uncaught exceptions this much. We could also handle the exceptions in the normal flow.
- In dev mode, since we want to keep the app running, a "normal" exception like a failed API call will recover after the uncaught exception is handled and continue the program where it left. This is very surprising.

In this pull request the error handling is changed a bit so that exceptions that are thrown in a go block are caught and put to the channel. A convenience macro `go-try` is implemented for that.