Improve API error message
Improve API error messages and formatting.

Biggest improvement is catching 401 error and rendering helpful error message with information about the marketplace and API key.

This PR also introduces a new format for error messages (small red arrow in front of each line). If we want we can start using this format for other errors as well.