Make sure that the key doesn't include dots.
Make sure that the key doesn't include dots. This underlines the fact that only top-level keys can be indexed.

However, this creates some inconsistency. Console allows adding new fields with keys that have dots, so there might be a top-level key with a dot in it. That can't now be indexed with the CLI. We can't really disallow dots in the key names altogether because JSON keys can contain dots.